moved makeScanDir from run_scans.py to utils/scanUtils.py so it could be imported
